+++ servers/maze_main_server/process/family/process.go
@@ -449,7 +449,7 @@ func (f *Family) OnConfirmApplyFamilyRQ_10589_10590(s *session.Session, req *Maz
 			return err
 		}
 		//加入家族群聊
-		err = groupservice.Default.InviteMember(ctx, app.Maze, familyInfo.FamilyGroupID, req.GetApplyUser().GetUserId())
+		err = groupservice.Default.InviteMember(ctx, app.Maze, int64(familyInfo.FamilyGroupID), req.GetApplyUser().GetUserId())
 		if err != nil {
 			res.ErrInfo = errors.MODULE_ERROR.Wrap("加入家族群聊失败")
 			logger.CtxError(ctx, "加入家族群聊失败", zap.Error(err))
@@ -462,7 +462,7 @@ func (f *Family) OnConfirmApplyFamilyRQ_10589_10590(s *session.Session, req *Maz
 			res.ErrInfo = errors.MODULE_ERROR.Wrap("获取联盟ID失败")
 			return err
 		}
-		err = groupservice.Default.InviteMember(ctx, app.Maze, allianceID, req.GetApplyUser().GetUserId())
+		err = groupservice.Default.InviteMember(ctx, app.Maze, int64(allianceID), req.GetApplyUser().GetUserId())
 		if err != nil {
 			res.ErrInfo = errors.MODULE_ERROR.Wrap("加入联盟群聊失败")
 			logger.CtxError(ctx, "加入联盟群聊失败", zap.Error(err))
